+++ CHANGES.rst
@@ -2,8 +2,12 @@
  Changes
 =========
 
-4.5 (unreleased)
+5.0 (unreleased)
 ================
+
+- Drop support for Python 2.7, 3.5, 3.6.
+
+- Drop support for deprecated ``python setup.py test``.
 
 - Add support for Python 3.11.